svc-composition-service/requirements.txt
requests==2.22.0
urllib3==1.25.3
six>=1.12.0
pyinstaller>=3.5
redis==3.3.8
setuptools==41.4.0
grpcio>=1.27.1
grpcio-tools>=1.27.1
protobuf>=3.6.1
tenacity>=6.1.0
etcd3>=0.12.0